pyproject.toml
[project]
name = "prepmd"
version = "0.1"
requires-python = ">= 3.13.3"
authors = [{name = "Rob Welch", email = "robert.welch@stfc.ac.uk"}]
description = "Automatically prepare pdb/mmCif structures for molecular dynamics simulations"
dependencies = [
  "requests",
  "openmm==8.2.0",
  "pdbfixer==1.11",
  "gemmi==0.7.1",
  #"modeller==10.7",
  "biopython==1.85",
]
readme = "README.md"
license = "AGPL-3.0-or-later"

[project.scripts]
prepmd = "prepmd.prepmd:entry_point"
